+++ src/Components/Components.js
@@ -1,7 +1,7 @@
 import Optimizer from "./Optimizer/Optimizer";
 import Planner from "./Planner/Planner";
 import Directions from "./Directions/Directions"
-import ProtectedRoute from "../Common/Services/ProtectedRoute";
+import ProtectedRoute from "./ProtectedRoute/ProtectedRoute";
 import AuthRoute from "../Common/Services/AuthRoute";
 import Auth from "./Auth/Auth";
 
@@ -18,13 +18,13 @@ export default function Components() {
     return (
         <Router>
             <Routes>
-                <Route path="/" element={<ProtectedRoute element={Planner} />} />
-                <Route path="/auth" element={<AuthRoute element={Auth} />} />
-                <Route path="/optimizer" element={<ProtectedRoute element={Optimizer} />} />
+                <Route path="/" element={<ProtectedRoute path="/" element={Planner} />} />
+                <Route path="/auth" element={<AuthRoute path="/auth" element={Auth} />} />
+                <Route path="/optimizer" element={<ProtectedRoute path="/optimizer" element={Optimizer} />} />
                 <Route path="/directions" element={<Directions />} />	{/* Routing */}	
-        		<Route path="/register" element={<AuthRoute element={Register} />} />
-                <Route path="/login" element={<AuthRoute element={Login} />} />
-        		<Route path="*" element={<Navigate to="/auth" replace />} />
+        		<Route path="/register" element={<AuthRoute path="register" element={Register} />} />
+                <Route path="/login" element={<AuthRoute path="login" element={Login} />} />
+        		<Route path="*" element={<Navigate to="/" replace />} />
                 {/* In the future, we may add aditional pages, such as a settings/preferences page */}
             </Routes>
         </Router>
